requirements.txt
langchain
langchain-openai
langgraph
redis
chromadb
pydantic
python-dotenv
tiktoken
pandas
tabulate
